requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==6.0.1
django-allauth==65.13.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
frozenlist==1.8.0
gunicorn==25.0.3
h11==0.16.0
hyperlink==21.0.0
hypothesis==6.150.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
kombu==5.6.2
livekit-api==1.1.0
livekit-protocol==1.1.2
msgpack==1.1.2
multidict==6.7.1
packaging==25.0
pillow==12.1.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
service-identity==24.2.0
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
types-protobuf==6.32.1.20251210
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
yarl==1.22.0
zope.interface==8.2
